+++ src/Client/ClientCore/Entities/ClientEntityBuilder.cs
@@ -56,12 +56,14 @@ public sealed class ClientEntityBuilder : AbstractEntityBuilder
         ManaComponentCollection manas,
         StatsComponentCollection stats,
         EquipmentTypeComponentCollection equipmentTypes,
+        LevelComponentCollection levels,
+        ExperienceComponentCollection experiences,
         EntityTable entityTable)
         : base(entityId, isLoad, entityManager, kinematics, blockTiles,
             aboveBlocks, playerCharacterTags, names, parents, drawables, animatedSprites,
             orientations, admins, blockPositions, castBlockShadows, pointLightSources, physics, boundingBoxes,
             castShadows, entityTypes, serverOnly, stackables, quantities, itemUses, npcFlags, useRanges,
-            healths, staminas, manas, stats, equipmentTypes, entityTable)
+            healths, staminas, manas, stats, equipmentTypes, levels, experiences, entityTable)
     {
     }