+++ src/Server/ServerCore/Entities/ServerEntityBuilder.cs
@@ -60,12 +60,14 @@ public sealed class ServerEntityBuilder : AbstractEntityBuilder
         ManaComponentCollection manas,
         StatsComponentCollection stats,
         EquipmentTypeComponentCollection equipmentTypes,
+        LevelComponentCollection levels,
+        ExperienceComponentCollection experiences,
         EntityTable entityTable)
         : base(entityId, load, entityManager, kinematics, blockTiles,
             aboveBlocks, playerCharacterTags, names, parents,
             drawables, animatedSprites, orientations, admins, blockPositions, castBlockShadows, pointLightSources,
             physics, boundingBoxes, castShadows, entityTypes, serverOnly, stackables, quantities, itemUses, npcFlags,
-            useRanges, healths, staminas, manas, stats, equipmentTypes, entityTable)
+            useRanges, healths, staminas, manas, stats, equipmentTypes, levels, experiences, entityTable)
     {
         this.accounts = accounts;
     }