+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### September
 
+#### 13 September 2026
+
+* Add new common, scriptable `Level` and `Experience` components for players and NPCs. The components are exposed to scripts as `Components.Level` and `Components.Experience` and as `Level` and `Experience` keys in the `Entities.Create` spec table, may be set on NPC templates in the template editor, and are persisted to the database. New players are created with `Level 1` and `Experience 0`.
+
 #### 12 September 2026
 
 * Add equipment and equippable items. Each new player is assigned one `EquipmentSlot` child entity per equipment type, and a new common, scriptable `EquipmentType` component (with the new `EquipmentType` value type: `Weapon`, `Offhand`, `Helmet`, `Chest`, `Leggings`, `Necklace`, `LeftRing`, `RightRing`) marks equippable items and configures equipment slots. The component is exposed to scripts as `Components.EquipmentType` and as an `EquipmentType` key in the `Entities.Create` spec table, may be set on item templates in the template editor, and is persisted to the database.